feat: CWA-2026-003
